add：爬虫任务保存

+++ orm/orm-novel/src/main/java/com/alice/novel/module/common/entity/ReptileDetailInfo.java
@@ -38,7 +38,7 @@ public class ReptileDetailInfo extends Model<ReptileDetailInfo> {
      * 爬虫任务主键
      */
     @TableField("REPTILE_INFO_ID")
-    private Integer reptileInfoId;
+    private Long reptileInfoId;
 
     /**
      * 爬虫任务明细地址
